refactor check oz page
