+++ src/components/CheckOZ/index.js
@@ -1,0 +1,4 @@
+export { default as AddressTabContent } from './AddressTabContent';
+export { default as CoordinatesTabContent } from './CoordinatesTabContent';
+export { default as ResultsDisplay } from './ResultsDisplay';
+export { default as ErrorDisplay } from './ErrorDisplay';
